requirements.txt
datasets>=2.18.0
pytest>=8.0.0
python-dotenv>=1.0.0
groq>=0.9.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
